Bug fix + no download of file if already present

+++ ClassDynSpecMS.py
@@ -39,10 +39,11 @@ class ClassDynSpecMS():
 
         if FileCoords is None:
             FileCoords="Transient_LOTTS.csv"
-            ssExec="wget -q --user=anonymous ftp://ftp.strw.leidenuniv.nl/pub/tasse/%s -O %s"%(FileCoords,FileCoords)
-            print>>log,"Downloading %s"%FileCoords
-            print>>log, "   Executing: %s"%ssExec
-            os.system(ssExec)
+            if not os.path.isfile(FileCoords):
+                ssExec="wget -q --user=anonymous ftp://ftp.strw.leidenuniv.nl/pub/tasse/%s -O %s"%(FileCoords,FileCoords)
+                print>>log,"Downloading %s"%FileCoords
+                print>>log, "   Executing: %s"%ssExec
+                os.system(ssExec)
         self.PosArray=np.genfromtxt(FileCoords,dtype=[('Name','S200'),("ra",np.float64),("dec",np.float64),('Type','S200')],delimiter=",")
         self.PosArray=self.PosArray.view(np.recarray)
         self.PosArray.ra*=np.pi/180.
